requirements.txt
# Face Tracking System Requirements

# Core Computer Vision
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0
Pillow>=9.5.0

# Face Recognition
insightface>=0.7.3
onnxruntime>=1.15.0

# Machine Learning & Tracking
scikit-learn>=1.3.0
filterpy>=1.4.5
lap>=0.5.0

# Web Dashboard (Bonus Feature)
Flask>=2.3.0
Flask-SocketIO>=5.3.0
python-socketio>=5.8.0
python-engineio>=4.7.0

# Database
sqlite3  # Built-in with Python

# Logging & Configuration
logging  # Built-in with Python
json     # Built-in with Python
pathlib  # Built-in with Python
datetime # Built-in with Python

# System & Threading
threading # Built-in with Python
argparse  # Built-in with Python
signal    # Built-in with Python
sys       # Built-in with Python
os        # Built-in with Python
time      # Built-in with Python
dataclasses  # Built-in with Python (Python 3.7+)
typing  # Built-in with Python

# Optional: For better performance
# torch==2.0.1
# torchvision==0.15.2

# Optional: For advanced tracking (if implementing DeepSort/ByteTrack)
# filterpy==1.4.5
# lap==0.4.0

# Optional: For web interface (future enhancement)
# flask==2.3.3
# flask-cors==4.0.0

# Development and Testing
# pytest==7.4.0
# pytest-cov==4.1.0
